requirements.txt
Flask==1.1.2
Jinja2==2.11.2
python-dotenv==0.15.0
tweepy==3.9.0